send linear acceleration only for bno055, and send raw int16_t values

+++ gc2_firmware/data_collection/gc_data.h
@@ -37,7 +37,7 @@ private:
   void read_battery_charge();
   bool need_report_battery_charge();
   float get_gyro_max();
-  void get_accel_1(float *accel_values);
+  void get_accel_1(int16_t *accel_values);
   void get_accel_2(float *accel_values);
   uint16_t read_emg();
   void emg_beep(uint16_t emg_value);
